move web standards rules to separate files and update build process to concat rules and dashboard code
