+++ Plugins/samples/index.html
@@ -6,7 +6,9 @@
     <link rel="stylesheet" href="index.css" />
     <script src="http://localhost:1337/vorlon.max.js"></script>
     <style>
-      
+      #webstandards{
+        transform : scale(1.1);
+      }
     </style>
   </head>
   <body>
@@ -68,10 +70,16 @@
         </p>
     </div>
     
+    <h2>web standards and best practices</h2>
+    <p>This block is for testing several aspects of web standards and best practice checks</p>
+    <p id="webstandards">
+      <object></object>
+    </p>
+
     <div id="footer">
       Learn more about <a href="http://vorlonjs.com">Vorlon.JS</a> or find the source at <a href="http://github.com/MicrosoftDX/Vorlonjs">GitHub</a>.
     </div>
-
+    
     <script type="text/javascript">
         console.log('The sample page has loaded!');
         function intenseLog() {
